move footer files into footer component folder
